fix(backend): custom_order 族 customer_id/product_id/color_id/sales_order_id 实库 INT8 → 模型回退 i64
- 撤销 98663d5a 对 custom_order 族的盲改:这 4 列经 system/mod.rs ALTER 为 BIGINT、
  sales_crm/mod.rs 以裸 BIGINT ADD COLUMN 创建,不经 fix_fk_types 改写,真实列宽 INT8;
  CI #4642 运行时证据(POST /custom-orders → decoding column "customer_id" mismatch)为终裁
- 回退范围: custom_order::Model、CreateCustomOrderDto、CustomOrderListItem、CustomOrderDetail
- 保持 i32 不动(真实 INT4,由 fix_fk_types 把 BIGINT REFERENCES <白名单表> 改写为 INTEGER):
  process_nodes.operator_id、process_logs.operator_id、after_sales.customer_id、
  AdvanceRequest.operator_id、UpdateProcessNodeDto/AdvanceNodeDto/AddProcessLogDto.operator_id、
  custom_order_state_service 的 operator_id 参数
- cargo fmt --check 通过;同型范式参照 e0647143 对 sales_quotation 族的修复

+++ backend/src/models/custom_order_create_dto.rs
@@ -11,9 +11,9 @@ use validator::Validate;
 /// 创建定制订单请求 DTO
 #[derive(Debug, Deserialize, Serialize, Validate, Clone)]
 pub struct CreateCustomOrderDto {
-    pub customer_id: i32,
-    pub product_id: i32,
-    pub color_id: Option<i32>,
+    pub customer_id: i64,
+    pub product_id: i64,
+    pub color_id: Option<i64>,
 
     /// 规格
     #[validate(length(min = 1, max = 200))]
@@ -42,7 +42,7 @@ pub struct CreateCustomOrderDto {
     pub expected_delivery_date: Option<NaiveDate>,
 
     /// 关联销售订单 ID（从销售订单转定制订单时使用）
-    pub sales_order_id: Option<i32>,
+    pub sales_order_id: Option<i64>,
 
     /// 金额
     pub total_amount: Option<Decimal>,
